requirements.txt
opencv-python
flask
numpy
torch
pillow
supervision